+++ src/librustc_privacy/lib.rs
@@ -9,8 +9,7 @@ extern crate syntax;
 use rustc::bug;
 use rustc::hir::def::{DefKind, Res};
 use rustc::hir::def_id::{CrateNum, DefId, CRATE_DEF_INDEX, LOCAL_CRATE};
-use rustc::hir::intravisit::{self, NestedVisitorMap, Visitor};
-use rustc::hir::itemlikevisit::DeepVisitor;
+use rustc::hir::intravisit::{self, DeepVisitor, NestedVisitorMap, Visitor};
 use rustc::hir::{self, AssocItemKind, HirIdSet, Node, PatKind};
 use rustc::lint;
 use rustc::middle::privacy::{AccessLevel, AccessLevels};
